started implementing C-side of aiFileIO
to properly import models from the already embedded models I need to
create a custom aiFileIO that reads the embed.FS and passes it back to
assimp. Due to implementation details of assimp, this needs to be
written in C rather than go.

aiFileIO is implemented and the skeleton of aiFile is made