mostly fixed animations
- verticies weren't being written back into the vertex list when
processing the mesh. Each vertex was written to and then discarded
rather than updating the corresponding version in the verticies slice

- vertex weights were getting the vertexId written to them rather than
the weight. A stupid typo that caused so many headaches

Something is still very off with the animation as there is a visual
problem (debugging shaders shows that the second boneId is always set
to 0 which is likely throwing something off) but it is now being shown
and there is some level of movement.