removed flipVertical from model.go
loading the textures and manually flipping them to the correct
orientation is costly especially when assimp can flip UVs for us.

Saves 440ms on launch time

had to change a couple textures to use 8 bit color precision rather
than 16 bit color precision because it's too much effort to write code
to convert all images to the right format for me.