fixed inverted camera
camera view matrix was being calculated backwards so camera position
was always negative of world space position

weird phenomenom where I'm not moving the camera around the world but
instead the world around the camera. Just a quirk of OpenGL ig.

this meant that the world should move negative of the camera position
to show the perception of the camera moving. E.G in a vacuum, moving you
left would look the same as me moving right